Implement functionality for the start button
The app defines a useRef and it is intially null. When the form component loads and it assigns the useRef passed down to one of the input fields, the useRef in the parent now has a pointer to that.

So when the button is clicked in the Hero component and it runs a function in the parent it can now apply focus to the element in the form.