Fix accessibility issues
Increased the size of the smallest font (min size should be 11px)

Linked the labels to the fields

Removed the button from the on hover section but kept them focusable

Removed icons from footer. Accessibility websites were saying the icons were too close together to be clickable for touch screens.