More accessibility fixes
The described by doesn't work because the tooltip isn't shown by default. So add the hovered data as a hidden description that we can then use with the described by. Although made use of the hidden styles to add an h1 for screen readers.

Added a role of none to the ClearIcon to let screen readers know its for decoration.

Changed the h4 to an h3 so that the headers are in order.