Pass current path
When the dashboard is clicked it is redirected to the Login page in here if the user isn't currently logged in.

After they log in we want the user to be taken to the dashboard page so we need to pass in the path the user selected to the Login componnent.