Rework links
Need to have a router link inside a mui link because we want the router functionality with the ability to style a mui link.

Also redirect to the path passed in if set, otherwise go to home i.e. /

replace: true - this is needed for the following reason. If I try to go to the dashboard and it redirects me to the log in. I then log it and it takes me to the dashboard. If the user hits the back button they will be taken back to the login page when they were only redirected there. So the replace: true replaces the log in page in the history stack with the dashboard entry. So when the user goes to the dashboard and presses the back button they will go back to the home page and not the login. That is a better user experience.