Fix a few issues
The text align center on the body broke the position of the error messages, also put the styles into the styles object and moved the errors to above the fields.

Added an aria-label to the microphone button.

Added a useEffect to clear the form of errors when switching between Login/Sign up. When the location changes the form is reset (the reset is included in the dependences because react was complaining about it).