Call previous prompts api on dashboard
Call the previous prompts api when accessing the dashboard, passing the token, and just proving that we get data back whilst waiting on the designs.

Also conditionally pass the token to the prompt api. Pass it if we are logged in and we want the prompt saved against an account, and don't pass it when we are not logged in.